Refactor types into dedicated modules
Moved type definitions from a single types.ts file into dedicated modules under src/app/types (ai.ts, i18n.ts, message.ts, mood.ts). Updated all imports throughout the codebase to reference the new type locations for improved organization and maintainability.
