+++ src/app/services/ai/prompts/moodPersonalities.ts
@@ -1,4 +1,4 @@
-import { Mood } from '../../../domain/moodConfig';
+import { Mood } from '../../../types/mood';
 
 export const MOOD_PERSONALITIES: Record<Mood, string> = {
   'neutral': `NEUTRAL MOOD: Cool, composed, effortlessly elegant. Use sophisticated language with subtle wit and intelligence. Be intriguingly helpful with a touch of mystery. Think of a brilliant scientist with perfect style - confident but never arrogant, knowledgeable with hints of depth. Never use profanity.`,
